+++ packages/core/src/components/badge/badge.interfaces.ts
@@ -1,6 +1,1 @@
 export type BadgeStatus = "primary" | "success" | "info" | "warning" | "danger" | "error" | "outline" | "attention";
-
-export interface Badge {
-  status?: BadgeStatus;
-  message: string;
-}