+++ packages/dso-toolkit/src/components/accordion/accordion.models.ts
@@ -1,3 +1,5 @@
+import { BadgeStatus } from "../badge";
+
 export type AccordionHandleElement = "anchor" | "button";
 
 export type AccordionVariant = "default" | "compact" | "conclusion" | "neutral" | "compact-black";
@@ -22,6 +24,8 @@ export interface AccordionSection<TemplateFnReturnType> {
   icon?: string;
   attachmentCount?: number;
   content?: TemplateFnReturnType;
+  badgeStatus?: BadgeStatus;
+  badgeMessage?: string;
   dsoToggleClick?: (e: CustomEvent<AccordionSectionToggleClickEvent>) => void;
   dsoAnimationStart?: (e: CustomEvent<AccordionSectionAnimationStartEvent>) => void;
   dsoAnimationEnd?: (e: CustomEvent<AccordionSectionAnimationEndEvent>) => void;