+++ storybook/src/components/accordion/accordion.css-template.ts
@@ -7,7 +7,7 @@ import { ComponentImplementation } from "../../templates";
 export const cssAccordion: ComponentImplementation<Accordion<TemplateResult>> = {
   component: "accordion",
   implementation: "html-css",
-  template: ({ attachmentsCounterTemplate, iconTemplate }) =>
+  template: ({ attachmentsCounterTemplate, iconTemplate, badgeTemplate }) =>
     function accordionTemplate(accordion) {
       const statusMap = new Map<string, string>([
         ["success", "succes"],
@@ -27,6 +27,9 @@ export const cssAccordion: ComponentImplementation<Accordion<TemplateResult>> = 
             ? html`<span class="dso-icon">${iconTemplate({ icon: section.icon })}</span>`
             : nothing}
           ${section.handleTitle}
+          ${section.badgeMessage
+            ? badgeTemplate({ status: section.badgeStatus, message: section.badgeMessage })
+            : nothing}
           ${section.icon && !accordion.reverseAlign
             ? html`<span class="dso-icon">${iconTemplate({ icon: section.icon })}</span>`
             : nothing}